requirements.txt
fastapi
interpreters-pep-734
pandas
scipy
sqlalchemy
sympy
